fix: move android name over
