+++ src-tauri/gen/android/app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.open_grind" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <style name="Theme.free_grind" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Customize your theme here. -->
     </style>
 </resources>
